# packages/schemas/openapi.yaml
components:
  schemas:
...
    CopyTemplateRequest:
      properties:
        name:
          anyOf:
            - type: string
            - type: 'null'
          title: Name
        source_ref:
          title: Source Ref
          type: string
      required:
        - source_ref
      title: CopyTemplateRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    PreviewRequest:
      properties:
        brand_color_hex:
          default: '#C4513A'
          pattern: ^#[0-9A-Fa-f]{6}$
          title: Brand Color Hex
          type: string
        copy:
          default: 新品上市
          title: Copy
          type: string
        design_note:
          default: keep product centered with premium spacing
          title: Design Note
          type: string
        locale:
          enum:
            - zh
            - en
          title: Locale
          type: string
        sample_brand:
          default: 示例品牌
          title: Sample Brand
          type: string
        sample_category:
          default: 服饰
          title: Sample Category
          type: string
        sample_name:
          default: 示例商品
          title: Sample Name
          type: string
        style_prompt:
          default: warm minimalist studio, soft daylight
          title: Style Prompt
          type: string
        template_ref:
          title: Template Ref
          type: string
        visual:
          default: single product hero image, clean ecommerce composition
          title: Visual
          type: string
      required:
        - template_ref
        - locale
      title: PreviewRequest
      type: object
    PreviewResponse:
      properties:
        cost_usd:
          title: Cost Usd
          type: number
        png_path:
          anyOf:
            - type: string
            - type: 'null'
          title: Png Path
        prompt:
          title: Prompt
          type: string
      required:
        - prompt
        - png_path
        - cost_usd
      title: PreviewResponse
      type: object
...
    SchemePayload:
      properties:
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        enabled:
          default: true
          title: Enabled
          type: boolean
        locale:
          enum:
            - zh
            - en
          title: Locale
          type: string
        name:
          minLength: 1
          title: Name
          type: string
        slots:
          items:
            $ref: '#/components/schemas/SchemeSlot'
          maxItems: 14
          minItems: 14
          title: Slots
          type: array
      required:
        - locale
        - name
        - slots
      title: SchemePayload
      type: object
...
    SchemeSummary:
      properties:
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        editable:
          default: true
          title: Editable
          type: boolean
        enabled:
          default: true
          title: Enabled
          type: boolean
        id:
          title: Id
          type: string
        locale:
          enum:
            - zh
            - en
          title: Locale
          type: string
        name:
          title: Name
          type: string
        slots:
          items:
            $ref: '#/components/schemas/SchemeSlot'
          title: Slots
          type: array
        source:
          default: custom
          enum:
            - built_in
            - custom
          title: Source
          type: string
      required:
        - id
        - name
        - locale
        - slots
      title: SchemeSummary
      type: object
...
    TemplatePayload:
      properties:
        category:
          default: lifestyle
          enum:
            - hero
            - detail_m3
            - lifestyle
            - short_video
            - amazon_hero
          title: Category
          type: string
        category_tips:
          additionalProperties:
            type: string
          title: Category Tips
          type: object
        defaults:
          additionalProperties:
            type: string
          title: Defaults
          type: object
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        enabled:
          default: true
          title: Enabled
          type: boolean
        examples:
          items:
            type: string
          title: Examples
          type: array
        locale:
          enum:
            - zh
            - en
          title: Locale
          type: string
        name:
          minLength: 1
          title: Name
          type: string
        prompt_template:
          additionalProperties:
            type: string
          minProperties: 1
          title: Prompt Template
          type: object
        supports_image_reference:
          default: false
          title: Supports Image Reference
          type: boolean
        tags:
          items:
            type: string
          title: Tags
          type: array
        variants:
          additionalProperties: true
          title: Variants
          type: object
      required:
        - locale
        - name
        - prompt_template
      title: TemplatePayload
      type: object
    TemplateSummary:
      properties:
        category:
          enum:
            - hero
            - detail_m3
            - lifestyle
            - short_video
            - amazon_hero
          title: Category
          type: string
        copyable:
          default: true
          title: Copyable
          type: boolean
        defaults:
          anyOf:
            - additionalProperties:
                type: string
              type: object
            - type: 'null'
          title: Defaults
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        editable:
          default: false
          title: Editable
          type: boolean
        enabled:
          default: true
          title: Enabled
          type: boolean
        examples:
          items:
            type: string
          title: Examples
          type: array
        id:
          title: Id
          type: string
        locale:
          enum:
            - zh
            - en
          title: Locale
          type: string
        name:
          title: Name
          type: string
        name_en:
          anyOf:
            - type: string
            - type: 'null'
          title: Name En
        prompt_template:
          anyOf:
            - additionalProperties:
                type: string
              type: object
            - type: 'null'
          title: Prompt Template
        source:
          default: built_in
          enum:
            - built_in
            - custom
          title: Source
          type: string
        tags:
          items:
            type: string
          title: Tags
          type: array
        thumbnail_url:
          anyOf:
            - type: string
            - type: 'null'
          title: Thumbnail Url
      required:
        - id
        - name
        - category
        - tags
        - locale
        - description
        - thumbnail_url
      title: TemplateSummary
      type: object
    TemplateUpdate:
      properties:
        category:
          anyOf:
            - enum:
                - hero
                - detail_m3
                - lifestyle
                - short_video
                - amazon_hero
              type: string
            - type: 'null'
          title: Category
        category_tips:
          anyOf:
            - additionalProperties:
                type: string
              type: object
            - type: 'null'
          title: Category Tips
        defaults:
          anyOf:
            - additionalProperties:
                type: string
              type: object
            - type: 'null'
          title: Defaults
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        enabled:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Enabled
        examples:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Examples
        name:
          anyOf:
            - type: string
            - type: 'null'
          title: Name
        prompt_template:
          anyOf:
            - additionalProperties:
                type: string
              type: object
            - type: 'null'
          title: Prompt Template
        supports_image_reference:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Supports Image Reference
        tags:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Tags
        variants:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Variants
      title: TemplateUpdate
      type: object
...
info:
  title: Viskit Studio API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/templates:
    get:
      operationId: get_templates_api_templates_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TemplateSummary'
                title: Response Get Templates Api Templates Get
                type: array
          description: Successful Response
      summary: Get Templates
      tags:
        - templates
    post:
      operationId: create_template_api_templates_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplatePayload'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateSummary'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Template
      tags:
        - templates
  /api/templates/copy:
    post:
      operationId: copy_template_api_templates_copy_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyTemplateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateSummary'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Copy Template
      tags:
        - templates
...
  /api/templates/preview:
    post:
      operationId: preview_template_api_templates_preview_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreviewRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreviewResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Preview Template
      tags:
        - templates
  /api/templates/schemes:
    get:
      operationId: list_schemes_api_templates_schemes_get
      parameters:
        - in: query
          name: locale
          required: false
          schema:
            default: zh
            enum:
              - zh
              - en
            title: Locale
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SchemeSummary'
                title: Response List Schemes Api Templates Schemes Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Schemes
      tags:
        - templates
    post:
      operationId: create_scheme_api_templates_schemes_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemePayload'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemeSummary'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Scheme
      tags:
        - templates
  /api/templates/{template_ref}:
    delete:
      operationId: delete_template_api_templates__template_ref__delete
      parameters:
        - in: path
          name: template_ref
          required: true
          schema:
            title: Template Ref
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties:
                  type: boolean
                title: Response Delete Template Api Templates  Template Ref  Delete
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Template
      tags:
        - templates
    patch:
      operationId: update_template_api_templates__template_ref__patch
      parameters:
        - in: path
          name: template_ref
          required: true
          schema:
            title: Template Ref
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateSummary'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Template
      tags:
        - templates
...